+++ packages/dockest/src/run/waitForServices/healthcheck.ts
@@ -1,41 +1,39 @@
+import { createDefaultHealthchecks } from '../../utils/createDefaultHealthchecks'
 import { DockestError } from '../../Errors'
-import { execaWrapper } from '../../utils/execaWrapper'
-import { sleep } from '../../utils/sleep'
 import { Runner } from '../../@types'
+import { sleep } from '../../utils/sleep'
 
 const logPrefix = '[Check Responsiveness]'
 
 export const checkResponsiveness = async (runner: Runner) => {
-  const {
-    containerId,
-    dockestService: { healthchecks = [] },
-    logger,
-  } = runner
+  const { containerId, dockerComposeFileService, healthcheck, logger } = runner
   const responsivenessTimeout = 30
 
-  await Promise.all(
-    healthchecks.map(async healthcheck => {
-      const recurse = async (responsivenessTimeout: number, runner: Runner) => {
-        logger.debug(`${logPrefix} Timeout in ${responsivenessTimeout}s`) // FIXME: Try getting replacePrevLine in here (may prove difficult if it runs in parallel)
+  const recurse = async (responsivenessTimeout: number, runner: Runner) => {
+    logger.debug(`${logPrefix} Timeout in ${responsivenessTimeout}s`) // FIXME: Try getting replacePrevLine in here (may prove difficult with `runInBand: false`)
 
-        if (responsivenessTimeout <= 0) {
-          throw new DockestError(`${logPrefix} Timed out`, { runner })
-        }
+    if (responsivenessTimeout <= 0) {
+      throw new DockestError(`${logPrefix} Timed out`, { runner })
+    }
 
-        try {
-          const command = healthcheck(runner.dockerComposeFileService, containerId)
-          await execaWrapper(command, { runner, logPrefix })
+    try {
+      logger.debug(`${logPrefix}`)
 
-          logger.info(`${logPrefix} Success`, { success: true })
-        } catch (error) {
-          responsivenessTimeout--
+      await healthcheck({
+        containerId,
+        defaultHealthchecks: createDefaultHealthchecks(runner),
+        dockerComposeFileService,
+        logger,
+      })
 
-          await sleep(1000)
-          await recurse(responsivenessTimeout, runner)
-        }
-      }
+      logger.info(`${logPrefix} Success`, { success: true })
+    } catch (error) {
+      responsivenessTimeout--
 
+      await sleep(1000)
       await recurse(responsivenessTimeout, runner)
-    }),
-  )
+    }
+  }
+
+  await recurse(responsivenessTimeout, runner)
 }