+++ packages/dockest/src/test-utils.ts
@@ -1,5 +1,6 @@
 import { DEFAULT_OPTS, DEFAULT_$ } from './constants'
-import { DockestConfig, DockestService, DockerComposeFile } from './@types'
+import { DockestConfig, DockestService, DockerComposeFile, Runner } from './@types'
+import { Logger } from './Logger'
 
 export const createConfig = (
   $: Partial<DockestConfig['$']> = {},
@@ -16,11 +17,23 @@ export const createConfig = (
   },
 })
 
+export const createRunner = (overrides?: Partial<Runner>): Runner => ({
+  commands: [],
+  containerId: '',
+  dependents: [],
+  dockerComposeFileService: { image: 'node:10-alpine', ports: [{ published: 3000, target: 3000 }] },
+  healthcheck: () => Promise.resolve(),
+  logger: new Logger('node'),
+  serviceName: 'node',
+  ...(overrides || {}),
+})
+
 export const DOCKEST_SERVICE: DockestService = {
   serviceName: 'redis',
 }
 
 export const DOCKER_COMPOSE_FILE: DockerComposeFile = {
+  version: '3.7',
   services: {
     [DOCKEST_SERVICE.serviceName]: {
       image: 'redis:5.0.3-alpine',
@@ -32,7 +45,6 @@ export const DOCKER_COMPOSE_FILE: DockerComposeFile = {
       ],
     },
   },
-  version: '3.7',
 }
 
 export const MERGED_COMPOSE_FILES = `