+++ packages/examples/node-to-node/dockest.ts
@@ -7,4 +7,15 @@ const { run } = new Dockest({
   composeFile: 'docker-compose.yml',
 })
 
-run([{ serviceName: 'node_to_node_users' }, { serviceName: 'node_to_node_orders' }])
+run([
+  {
+    serviceName: 'node_to_node_orders',
+    commands: ['echo "Hello from orders (dependency - should run first) 👋🏽"'],
+    dependents: [
+      {
+        serviceName: 'node_to_node_users',
+        commands: ['echo "Hello from users (dependent - should run right after orders) 👋🏽"'],
+      },
+    ],
+  },
+])